@@ -10,7 +10,7 @@ const MenuItem = ({ title, imageUrl, size, linkUrl, match }) => (
       style={{ backgroundImage: `url(${imageUrl})` }}
     />
     <div className='content'>
-      <h1 className='title'>{title.toUpperCase()}</h1>
+      <h1 className='title'>{title.toUpperCase()}</h1> 
       <span className='subtitle'>SHOP NOW</span>
     </div>
   </div>
